vpnrouting: further optimization, clean-up

+++ release/src-rt-6.x.4708/router/others/updown-client.sh
@@ -13,38 +13,13 @@ SERVICE=$(echo $dev | sed 's/\(tun\|tap\)1/client/;s/\(tun\|tap\)2/server/')
 DNSDIR="/etc/openvpn/dns"
 DNSCONFFILE="$DNSDIR/$SERVICE.conf"
 DNSRESOLVFILE="$DNSDIR/$SERVICE.resolv"
-RESTART_DNSMASQ=0
 FOREIGN_OPTIONS=$(set | grep "^foreign_option_" | sed "s/^\(.*\)=.*$/\1/g")
 LOGS="logger -t openvpn-updown-client.sh[$PID][$IFACE]"
+FILEEXISTS=""
 
-
-find_iface() {
-	if [ "$SERVICE" == "client1" ]; then
-		ID="311"
-	elif [ "$SERVICE" == "client2" ]; then
-		ID="312"
-	elif [ "$SERVICE" == "client3" ]; then
-		ID="313"
-	else
-		$LOGS "Interface not found!"
-		exit 0
-	fi
-}
 
 startAdns() {
-	local fileexists="" optionname option
-
-	[ "$(nvram get "vpn_"$SERVICE"_adns")" -eq 0 ] && return
-
-	[ ! -d $DNSDIR ] && mkdir $DNSDIR
-	[ -f $DNSCONFFILE ] && {
-		rm $DNSCONFFILE
-		fileexists=1
-	}
-	[ -f $DNSRESOLVFILE ] && {
-		rm $DNSRESOLVFILE
-		fileexists=1
-	}
+	local optionname option
 
 	[ -n "$FOREIGN_OPTIONS" ] & {
 		$LOGS "FOREIGN_OPTIONS: $FOREIGN_OPTIONS"
@@ -61,34 +36,28 @@ startAdns() {
 	[ ! -f $DNSCONFFILE -a ! -f $DNSRESOLVFILE ] && {
 		$LOGS "Warning: 'Accept DNS configuration' is enabled but no foreign options (push dhcp-option) have been received from the OpenVPN server!"
 	}
-
-	[ -f $DNSCONFFILE -o -f $DNSRESOLVFILE -o -n "$fileexists" ] && RESTART_DNSMASQ=1
 }
 
 stopAdns() {
-	local fileexists=""
-
-	[ -f $DNSCONFFILE ] && {
-		rm $DNSCONFFILE
-		fileexists=1
-	}
-	[ -f $DNSRESOLVFILE ] && {
-		rm $DNSRESOLVFILE
-		fileexists=1
-	}
-
-	[ -n "$fileexists" ] && RESTART_DNSMASQ=1
-}
-
-checkRestart() {
-	[ "$RESTART_DNSMASQ" -eq 1 ] && service dnsmasq restart
+	# nothing here, for now...
+	return
 }
 
 
 ###################################################
 
 
-find_iface
+[ "$SERVICE" == "" ] || [ "$(nvram get "vpn_"$SERVICE"_adns")" -eq 0 ] && exit 0
+
+[ ! -d $DNSDIR ] && mkdir $DNSDIR
+[ -f $DNSCONFFILE ] && {
+	rm $DNSCONFFILE
+	FILEEXISTS=1
+}
+[ -f $DNSRESOLVFILE ] && {
+	rm $DNSRESOLVFILE
+	FILEEXISTS=1
+}
 
 [ "$script_type" == "up" ] && {
 	startAdns
@@ -98,6 +67,8 @@ find_iface
 	stopAdns
 }
 
-checkRestart
+[ -f $DNSCONFFILE -o -f $DNSRESOLVFILE -o -n "$FILEEXISTS" ] && service dnsmasq restart
+
+rmdir -rf $DNSDIR
 
 exit 0
